added remote building of typst files via a post request + config and component information to select file

+++ src/label/label.rs
@@ -1,31 +1,45 @@
-use std::{fs, path::PathBuf};
+use std::{fs, path::{Path, PathBuf}, sync::Arc};
 
+use axum::extract::State;
 use typst::{foundations::{Dict, Value}, Library};
 use typst_pdf::PdfOptions;
 
-use crate::db::components::Component;
+use crate::{cli::config::Config, db::components::Component};
 
 use super::typst_wrapper;
 
 pub trait Label {
-    fn build(&self);
+
+    //fn debug_build(&self, label_path: String);
+    
+    fn build(&self, label_location: &str);
 
     fn get_inputs(&self) -> Library;
 }
 
 impl Label for Component{
-    fn build(&self) {
-        let data = fs::read_to_string("./test.typ").expect("Unable to read File!");
+    fn build(&self, label_location: &str) {
 
-        let world = typst_wrapper::TypstWrapperWorld::new("./".to_owned(), data, self.get_inputs());
+        if let Some(label) = &self.label {
+            let path = PathBuf::new().join(label_location).join(label.to_owned()+".typ");
+            //let path = Path::new(label_location).join(label.to_owned()+".typ");
+            
+            if fs::exists(&path).is_ok(){
+                let data = fs::read_to_string(path).expect("Unable to read File!");
 
-        let document: typst::layout::PagedDocument = typst::compile(&world)
-            .output
-            .expect("ahahha!");
+                let world = typst_wrapper::TypstWrapperWorld::new("./".to_owned(), data, self.get_inputs());
 
-        let pdf = typst_pdf::pdf(&document, &PdfOptions::default()).expect("ERROR EXPORTING");
+                let document: typst::layout::PagedDocument = typst::compile(&world)
+                    .output
+                    .expect("ahahha!");
 
-        fs::write("./output.pdf",pdf).expect("Error Writing");
+                let pdf = typst_pdf::pdf(&document, &PdfOptions::default()).expect("ERROR EXPORTING");
+
+                fs::write("./output.pdf",pdf).expect("Error Writing");
+            }
+        }
+
+        
     }
 
     fn get_inputs(&self) -> Library {
@@ -46,6 +60,8 @@ impl Label for Component{
 
         Library::builder().with_inputs(dict).build()
     }
+    
+    
 }
 
 fn insert_optional(dict: &mut Dict, key: &str, value: &Option<String>) {
